[indexer][generator] Moved tree_node from generator to indexer.
